CLOSE #26384 Implement STOCK_USE_WAREHOUSE_USAGE for use of non-counting warehouses e.g. quarantine
This feature implements the global variable of STOCK_USE_WAREHOUSE_USAGE. This variable allows warehouses to exclude their products from stock calculations. A button was added to the stock module in the admin panel. A simple "AND" check for the warehouse_usage column in the entrepot SQL table is added to the load_stock function to facilitate this change.