CLOSE #30347 Add functionality to select parts via lot and serial number in shipment creation
This commit adds the functionality to increment the qty to ship when creating a shipment via input of a lot or serial number. This option has to be activated in the shipments module. It works by enclosing the batch detail in a span with the id equal to the batch number. The JS script uses JQuery to relatively select the input box for "qty to ship" and increments the value.