Qual: Phan: Exclude htdocs/core/mymodule seen in cti.dolibarr.org
# Qual: Phan: Exclude htdocs/core/mymodule seen in cti.dolibarr.org

`htdocs/core/mymodule` has notices on https://cti.dolibarr.org
but it is not seen in git.

It would be a good idea to generate a module and include it in the
static analysis on cti, but `htdocs/core/mymodule` is not up-to-date.

If such a generation is done, a better name could be used.
For instance `generatedmymoduleforcti`.