NEW option: assign default roles to "individual" third-party contacts #30499
#30499

When the option to automatically create a contact for third parties of the "Individuals" type is active (THIRDPARTY_SUGGEST_ALSO_ADDRESS_CREATION).

Allow to define the default roles assigned to these contacts.

Process:

Selection of default roles in the settings (and saving)
Creation of a third party of the "Individuals" type and therefore of its associated contact.
The contact thus created is automatically assigned the predefined roles.