Qual: Update copyright alignment
# Qual: Update copyright alignment

One can observe that copyright statements already generated are manually
aligned to the shorter statements in the file.

32 characters (or 8 tabs) was considered the maximum width for the
name column in the entire codebase and used to generate the new
copyright notices.
However, this is often shortend.

This update makes the fixcopyrightheader.php check the copyright notices
already present in the file and adapt the new notice to those