FIX: Add missing list value hook for appended column
In the warehouse stock list, there is a 'printFieldPreListTitle' hook to prepend a new column in the table list and a 'printFieldListTitle' hook to append a new column in the table list. However, the 'printFieldListValue' hook shall be renamed as 'printFieldPreListValue' to populate any pre-pended columns whereas the hook named 'printFieldListValue' shall be moved at the end of the table to be consistent and be able to populate the appended column.