QUAL: AccountancySystem: maintenance work to uniformize and clean up (#31391)
* accountancysystem: inherit CommonObject

This provides the common methods on the database objects like error
handling.

* accountancysystem: improve error reporting

Forward the database error to the caller when an error happens.

* accountancysystem: clean up class

The fields that were completely unused are removed. $ref is still kept
since it was filled by AccountancySystem::fetch(), but marqued as
deprecated to incite using the newer more descriptive pcg_version field.

* Update accountancysystem.class.php

---------

Co-authored-by: Laurent Destailleur <eldy@destailleur.fr>