NEW Introduce prorata_discount and prorata_rate as new Facture object attributes
Building on an existing db field, introduce in Facture class a property prorata_discount (amount in currency) and prorata_rate (virtual property).
It also alter total_ht with a hook to update discount based on rate.
Prorata, as the name suggest, is a discount whose % is fixed and amount varies with total excl. taxes