CLOSE #33506 add a constant holding the ratio to resize images in ODT templates
A new Dolibarr constant 					MAIN_DOC_ODT_IMAGE_RATIO
is taken in consideration for ODT templates.
This constant holds a number that is multiplied when computing the size of images in ODT templates.
All doc_generic_*_odt_modules.php files have been updated to call the setImage Method with a new parameter $ratio
The method setImage in the file odt.php has been modified to have a new parameter $ratio and use it to compute the image size