linking member subscription contribution <-> proposal, order and invoice This can both be done manually from an existing proposal/order/invoice to a subscription created without creating an invoice. I use this when we sell memberships as part of registration for our membership classes.
The link will also be shown on the invoice that can be created as part of the subscription contribution creation.

These links show up in the json from the API on proposals, orders and invoices. The subscription does not show anything. It is currently unknown if the API can do this link or it will have to be done manually in the GUI.