Fix #35241: extrafields are not tranfered from order to shipping and cannot be edited at creation time
If there are the same extrafields in the customer order and the shipping, then the extrafields where not transfered to the order. Additionally extrafields could be edited on shipping creation time. This leads to a lock in case the extrafields is mandatory or could only be set on creation time.