FIX: Workflow for classify order as closed when shipping is validated or closed
With V21 this was changed the closure of the customer order was changed from setStatut to cloture. As cloture checks the permissions and setStatut nocht, this blocks the automatic workflow for cases where the user doesnt have the permission - but as this is an automatic workflow the permission of the user is not needed in my opinion.