FIX: global search, list view new search is not working
If you enter something in the global search with an input for example for products where no results are shown, you are in some kind of deadlock. You can enter new search values in the list view, but in the background search_all is added as hidden parameter which prevents the new search.